feat: add BikeAvailabilitySection component to manage and display bike availability, integrating it into CustomerDashboard and Home pages
