+++ frontend/src/pages/CustomerDashboard.tsx
@@ -1,67 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import CustomerHeader from '../components/CustomerHeader';
 import CustomerMessaging from '../components/messaging/CustomerMessaging';
+import BikeAvailabilitySection from '../components/BikeAvailabilitySection';
 import { heartbeatService } from '../Services/heartbeatService';
-import { bikeInventoryService } from '../Services/bikeInventoryService';
-import type { BikeAvailabilityResponse } from '../Services/bikeInventoryService';
-
-// Interface for the bike availability data
-interface BikeAvailability {
-  bikeType: string;
-  availableCount: number;
-  status: string;
-  pricing: {
-    minHourlyRate: number;
-    maxHourlyRate: number;
-    avgHourlyRate: number;
-  };
-  features: {
-    maxSpeed: string;
-    batteryLife: string;
-    weightCapacity: string;
-    specialFeatures: string[];
-  };
-  sampleBikes: any[];
-}
 
 const CustomerDashboard: React.FC = () => {
   const navigate = useNavigate();
   const [activeSection, setActiveSection] = useState<'dashboard' | 'messaging' | 'bikes'>('dashboard');
-  const [bikeAvailability, setBikeAvailability] = useState<BikeAvailability[]>([]);
-  const [totalAvailable, setTotalAvailable] = useState(0);
-  const [lastUpdated, setLastUpdated] = useState("");
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchBikes = async () => {
-      try {
-        setLoading(true);
-        const response = await bikeInventoryService.getAvailableBikes();
-        console.log(response);
-        if (response.success && response.bikeAvailability) {
-          setBikeAvailability(response.bikeAvailability);
-          setTotalAvailable(response.totalAvailable);
-          setLastUpdated(response.lastUpdated);
-        }
-      } catch (error) {
-        console.error("Error fetching available bikes:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchBikes();
-  }, []);
-
-  const formatDate = (dateString: string) => {
-    return new Date(dateString).toLocaleString('en-CA', {
-      month: 'short',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  };
 
   const handleLogout = () => {
     // Stop heartbeat service
@@ -164,7 +110,7 @@ const CustomerDashboard: React.FC = () => {
                     </div>
                     <div className="ml-4">
                       <p className="text-sm font-medium text-gray-600">Available Bikes</p>
-                      <p className="text-2xl font-bold text-gray-900">{totalAvailable}</p>
+                      <p className="text-2xl font-bold text-gray-900">-</p>
                     </div>
                   </div>
                 </div>
@@ -305,121 +251,7 @@ const CustomerDashboard: React.FC = () => {
           )}
 
           {activeSection === 'bikes' && (
-            <div>
-              <div className="mb-8">
-                <div className="flex justify-between items-center mb-4">
-                  <h3 className="text-2xl font-bold text-gray-900">Available Bikes</h3>
-                  <div className="text-sm text-gray-600">
-                    {totalAvailable} available • Updated {lastUpdated ? formatDate(lastUpdated) : 'N/A'}
-                  </div>
-                </div>
-              </div>
-
-              {loading ? (
-                <div className="flex justify-center items-center h-32">
-                  <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-amber-500"></div>
-                </div>
-              ) : (
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                  {bikeAvailability.length > 0 ? (
-                    bikeAvailability.map((bike, index) => (
-                      <div
-                        key={index}
-                        className={`bg-white rounded-xl shadow-sm border transition-all duration-200 hover:shadow-md ${
-                          bike.status === 'available' ? 'border-green-200' : 'border-gray-200'
-                        }`}
-                      >
-                        {/* Card Header */}
-                        <div className="p-6 border-b border-gray-100">
-                          <div className="flex justify-between items-start mb-3">
-                            <h4 className="text-lg font-semibold text-gray-900">{bike.bikeType}</h4>
-                            <span className={`px-3 py-1 rounded-full text-xs font-medium ${
-                              bike.status === 'available' 
-                                ? 'bg-amber-100 text-amber-800 border border-amber-300' 
-                                : 'bg-gray-100 text-gray-400 border border-gray-200'
-                            }`}>
-                              {bike.availableCount} available
-                            </span>
-                          </div>
-                          <div className="text-2xl font-bold text-amber-600">
-                            ${bike.pricing.avgHourlyRate}/hr
-                          </div>
-                        </div>
-
-                        {/* Features */}
-                        {bike.features.specialFeatures.length > 0 && (
-                          <div className="p-6 border-b border-gray-100">
-                            <div className="flex flex-wrap gap-2">
-                              {bike.features.specialFeatures.slice(0, 3).map((feature, featureIndex) => (
-                                <span
-                                  key={featureIndex}
-                                  className="px-2 py-1 bg-amber-100 text-amber-800 text-xs rounded-md border border-amber-200"
-                                >
-                                  {feature}
-                                </span>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-
-                        {/* Sample Bikes */}
-                        {bike.sampleBikes && bike.sampleBikes.length > 0 && (
-                          <div className="p-6">
-                            <p className="text-sm font-medium text-gray-700 mb-3">Available Bikes</p>
-                            <div className="space-y-2">
-                              {bike.sampleBikes.slice(0, 2).map((sampleBike, bikeIndex) => (
-                                <div
-                                  key={bikeIndex}
-                                  className="flex justify-between items-center p-3 bg-amber-50 rounded-lg border border-amber-100"
-                                >
-                                  <div>
-                                    <p className="text-sm font-medium text-gray-800">
-                                      {sampleBike.bikeId}
-                                    </p>
-                                    <p className="text-xs text-gray-600">
-                                      {sampleBike.location.address}
-                                    </p>
-                                  </div>
-                                  <div className="text-right">
-                                    <span className={`px-2 py-1 rounded-full text-xs font-medium border ${
-                                      sampleBike.status === 'available' 
-                                        ? 'bg-amber-200 text-amber-900 border-amber-300' 
-                                        : 'bg-gray-200 text-gray-400 border-gray-300'
-                                    }`}>
-                                      {sampleBike.status}
-                                    </span>
-                                  </div>
-                                </div>
-                              ))}
-                              {bike.sampleBikes.length > 2 && (
-                                <p className="text-xs text-amber-600 text-center">
-                                  +{bike.sampleBikes.length - 2} more bikes
-                                </p>
-                              )}
-                            </div>
-                          </div>
-                        )}
-
-                        {bike.sampleBikes && bike.sampleBikes.length === 0 && bike.status === 'available' && (
-                          <div className="p-6 text-center">
-                            <p className="text-sm text-gray-500">No bikes currently available</p>
-                          </div>
-                        )}
-                      </div>
-                    ))
-                  ) : (
-                    <div className="col-span-full text-center py-12">
-                      <div className="text-gray-400 mb-4">
-                        <svg className="mx-auto h-12 w-12" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
-                        </svg>
-                      </div>
-                      <p className="text-gray-500">No bike information available at this time.</p>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
+            <BikeAvailabilitySection title="Available Bikes" showHeader={true} />
           )}
 
           {activeSection === 'messaging' && (
